% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_IT_bm_mod.R
\docType{data}
\name{all_IT_bm_mod}
\alias{all_IT_bm_mod}
\title{A BRMS model fit estimating bee foraging ranges (m) by intertegular distance (mm), easurement type and phylogeny}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
all_IT_bm_mod
}
\description{
A BRMS model fit estimating bee foraging ranges (m) by intertegular distance (mm), easurement type and phylogeny
}
\keyword{datasets}
