% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_social_mod.R
\docType{data}
\name{all_social_mod}
\alias{all_social_mod}
\title{A BRMS model fit estimating bee foraging ranges (m) by intertegular distance (mm), sociality, measurement type and phylogeny.}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
all_social_mod
}
\description{
A BRMS model fit estimating bee foraging ranges (m) by intertegular distance (mm), sociality, measurement type and phylogeny.
}
\keyword{datasets}
