% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_IT.R
\docType{data}
\name{bee_IT}
\alias{bee_IT}
\title{Model fit Bee body size ~ IT + (1|Region/Species)}
\format{An object of class \code{brmsfit} of length 19.}
\usage{
bee_IT
}
\description{
Simplified bee body size model: Dry weight ~ ITD + (1|Region/Species)
}
\keyword{datasets}
