% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_phy_mod.R
\docType{data}
\name{bee_phy_mod}
\alias{bee_phy_mod}
\title{A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm), sex and phylogeny for ~400 bee species.}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
bee_phy_mod
}
\description{
A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm), sex and phylogeny for ~400 bee species.
}
\keyword{datasets}
