% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_sex_mod.R
\docType{data}
\name{bee_sex}
\alias{bee_sex}
\title{A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm) * sex for ~400 bee species.}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
bee_sex
}
\description{
A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm) * sex for ~400 bee species.
}
\keyword{datasets}
