% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_tax_mod.R
\docType{data}
\name{bee_tax_mod}
\alias{bee_tax_mod}
\title{Model fit Bee body size ~ ITD * Family + Sex +  (1|Region/Species)}
\format{An object of class \code{brmsfit} of length 19.}
\usage{
bee_tax_mod
}
\description{
Full taxonomic GLMM for estimating bee body size: Dry weight ~ ITD * Family + Sex +  (1|Region/Species).
}
\keyword{datasets}
