% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1.R
\docType{data}
\name{h1}
\alias{h1}
\title{Model fit object for estimating hoverfly body size ~ ITD + Sex + (1|Region/Species)}
\format{An object of class \code{brmsfit} of length 19.}
\usage{
h1
}
\description{
Hoverfly body size ~ ITD + Sex + (1|Region/Species)
}
\keyword{datasets}
