% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\docType{data}
\name{h2}
\alias{h2}
\title{Model fit object for estimating hoverfly body size ~ ITD + Sex + Subfamily + (1|Region/Species)}
\format{An object of class \code{brmsfit} of length 19.}
\usage{
h2
}
\description{
Hoverfly body size ~ ITD + Sex + Subfamily +  (1|Region/Species)
}
\keyword{datasets}
