% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hov_IT.R
\docType{data}
\name{hov_IT}
\alias{hov_IT}
\title{Model fit object: Hoverfly body size ~ IT + (1|Region/Species)}
\format{An object of class \code{brmsfit} of length 19.}
\usage{
hov_IT
}
\description{
Simplified hoverfly body size model: Bee body size ~ IT + (1|Region/Species)
}
\keyword{datasets}
