% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hov_sex_mod.R
\docType{data}
\name{hov_sex}
\alias{hov_sex}
\title{A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm) + sex for ~100 hoverfly species.}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
hov_sex
}
\description{
A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm) + sex for ~100 hoverfly species.
}
\keyword{datasets}
