% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hov_tax_mod.R
\docType{data}
\name{hov_tax_mod}
\alias{hov_tax_mod}
\title{A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm), sex and taxonomic subfamily for ~100 hoverfly species.}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
hov_tax_mod
}
\description{
A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm), sex and taxonomic subfamily for ~100 hoverfly species.
}
\keyword{datasets}
