% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_social_mod.R
\docType{data}
\name{max_social_mod}
\alias{max_social_mod}
\title{A BRMS model fit estimating maximum bee foraging range (m) by intertegular distance (mm), sociality, measurement type and phylogeny.}
\format{
An object of class \code{brmsfit} of length 21.
}
\usage{
max_social_mod
}
\description{
A BRMS model fit estimating maximum bee foraging range (m) by intertegular distance (mm), sociality, measurement type and phylogeny.
}
\keyword{datasets}
