% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_IT.R
\docType{data}
\name{bee_IT}
\alias{bee_IT}
\title{A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm) for ~400 bee species.}
\format{
An object of class \code{brmsfit} of length 19.
}
\usage{
bee_IT
}
\description{
A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm) for ~400 bee species.
}
\keyword{datasets}
