% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee_phy_mod.R
\docType{data}
\name{bee_phy_mod}
\alias{bee_phy_mod}
\title{Model fit Bee body size ~ ITD * Sex + (1|Region/Species)}
\format{An object of class \code{brmsfit} of length 19.}
\usage{
bee_phy_mod
}
\description{
Full phylogenetic GLMM for estimating bee body size: Dry weight ~ ITD * Sex + (1|Region/Species). Species is constrained by phylogenetic branch lengths
}
\keyword{datasets}
