% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hov_IT.R
\docType{data}
\name{hov_IT}
\alias{hov_IT}
\title{A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm) for ~100 hoverfly species.}
\format{An object of class \code{brmsfit} of length 19.}
\usage{
hov_IT
}
\description{
A BRMS model fit estimating the dry weight (mg) by intertegular distance (mm) for ~100 hoverfly species.
}
\keyword{datasets}
